refactor(iocuak): move isContainerModuleClassMetadataApi to calculations
